gv correction vm en cours. c long

+++ corewar_project/vm/srcs/ft_tools2.c
@@ -6,7 +6,7 @@
 /*   By: gvannest <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/12 15:12:33 by gvannest          #+#    #+#             */
-/*   Updated: 2018/07/26 14:40:06 by gvannest         ###   ########.fr       */
+/*   Updated: 2018/08/02 19:27:30 by gvannest         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,10 +42,10 @@ void		ft_update_map_pyr(int *map_pyr, int pc_act, int start, size_t k)
 
 void		ft_calc_index(int *k, int value, int pc_act, char flag_idx)
 {
-	unsigned int	mod;
+	short	mod;
 
-	mod = flag_idx == 1 ? IDX_MOD : MEM_SIZE;
-	*k = pc_act + (short int)value % mod;
+	mod = (flag_idx == 1 ? IDX_MOD : MEM_SIZE);
+	*k = pc_act + (short)value % mod;
 	(*k < 0 ? *k = *k + MEM_SIZE : *k);
 }
